+++ back/src/main/java/com/ll/goohaeyou/image/application/JobPostImageService.java
@@ -1,39 +1,28 @@
 package com.ll.goohaeyou.image.application;
 
-import com.ll.goohaeyou.auth.exception.AuthException;
-import com.ll.goohaeyou.global.exception.EntityNotFoundException;
+import com.ll.goohaeyou.image.domain.JobPostImageDomainService;
 import com.ll.goohaeyou.image.domain.policy.JobPostImagePolicy;
-import com.ll.goohaeyou.image.exception.ImageException;
-import com.ll.goohaeyou.jobApplication.domain.ImageStorage;
+import com.ll.goohaeyou.jobPost.jobPost.domain.JobPostDetailDomainService;
 import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPostDetail;
 import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPostImage;
-import com.ll.goohaeyou.jobPost.jobPost.domain.repository.JobPostDetailRepository;
-import com.ll.goohaeyou.jobPost.jobPost.domain.repository.JobPostImageRepository;
-import com.ll.goohaeyou.jobPost.jobPost.domain.repository.JobPostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class JobPostImageService {
-    private final JobPostRepository JobPostRepository;
-    private final ImageStorage imageStorage;
-    private final JobPostImageRepository jobPostImageRepository;
-    private final JobPostDetailRepository jobPostDetailRepository;
-    private final JobPostRepository jobPostRepository;
     private final JobPostImagePolicy jobPostImagePolicy;
+    private final JobPostDetailDomainService jobPostDetailDomainService;
+    private final JobPostImageDomainService jobPostImageDomainService;
 
     @Transactional
     public void uploadJobPostImages(String username, long postDetailId, MultipartFile[] jobPostImageFiles) {
-        JobPostDetail jobPostDetail = JobPostRepository.findById(postDetailId)
-                .orElseThrow(EntityNotFoundException.PostNotExistsException::new)
-                .getJobPostDetail();
+        JobPostDetail jobPostDetail = jobPostDetailDomainService.getById(postDetailId);
 
         jobPostImagePolicy.validateCanUploadImages(username, jobPostDetail, jobPostImageFiles);
 
@@ -41,29 +30,13 @@ public class JobPostImageService {
             deleteJobPostImages(username, postDetailId);
         }
 
-        List<JobPostImage> jobPostImages = new ArrayList<>();
-
-        boolean isMainNotSet = true;
-
-        for (MultipartFile imageFile : jobPostImageFiles) {
-            String imageUrl = imageStorage.upload(imageFile);
-
-            JobPostImage newJobPostImage = JobPostImage.create(imageUrl, isMainNotSet, jobPostDetail);
-
-            jobPostImages.add(newJobPostImage);
-            isMainNotSet = false;
-        }
-
-        jobPostDetail.getJobPostImages().addAll(jobPostImages);
-        jobPostDetailRepository.save(jobPostDetail);
+        jobPostImageDomainService.addJobPostImages(jobPostDetail, jobPostImageFiles);
     }
 
     public List<String> getJobPostImages(Long postDetailId) {
-        List<JobPostImage> postImages = jobPostImageRepository.findByJobPostDetailId(postDetailId);
+        List<JobPostImage> postImages = jobPostImageDomainService.getByJobPostDetailId(postDetailId);
 
-        if (postImages.isEmpty()) {
-            throw new ImageException.PostImagesNotFoundException();
-        }
+        jobPostImagePolicy.verifyPostImagesNotEmpty(postImages);
 
         return postImages.stream()
                 .map(JobPostImage::getJobPostImageUrl)
@@ -72,41 +45,20 @@ public class JobPostImageService {
 
     @Transactional
     public void deleteJobPostImages(String username, long postDetailId) {
-        JobPostDetail jobPostDetail = jobPostRepository.findById(postDetailId)
-                .orElseThrow(EntityNotFoundException.PostNotExistsException::new)
-                .getJobPostDetail();
+        JobPostDetail jobPostDetail = jobPostDetailDomainService.getById(postDetailId);
         List<JobPostImage> jobPostImages = jobPostDetail.getJobPostImages();
 
-        if (!jobPostDetail.getAuthor().equals(username)) {
-            throw new AuthException.NotAuthorizedException();
-        }
+        jobPostImagePolicy.verifyCanDeleteJobPostImages(username, jobPostDetail, jobPostImages);
 
-        if (jobPostImages.isEmpty()) {
-            throw new ImageException.PostImagesNotFoundException();
-        }
-
-        for (JobPostImage jobPostImage : jobPostImages) {
-            imageStorage.deleteImageFromS3(jobPostImage.getJobPostImageUrl());
-        }
-
-        jobPostImageRepository.deleteAll(jobPostImages);
-        jobPostDetail.getJobPostImages().clear();
+        jobPostImageDomainService.deleteByJobPostImages(jobPostImages, jobPostDetail);
     }
 
     @Transactional
     public void changeMainImage(String username, Long postId, Long currentImageId, Long newImageId) {
-        JobPostDetail jobPostDetail = jobPostDetailRepository.findById(postId)
-                .orElseThrow(EntityNotFoundException.PostNotExistsException::new);
+        JobPostDetail jobPostDetail = jobPostDetailDomainService.getById(postId);
 
         jobPostImagePolicy.verifyAuthor(username, jobPostDetail);
 
-        JobPostImage currentMainImage = jobPostImageRepository.findById(currentImageId)
-                .orElseThrow(ImageException.ImageNotFoundException::new);
-
-        JobPostImage newMainImage = jobPostImageRepository.findById(newImageId)
-                .orElseThrow(ImageException.ImageNotFoundException::new);
-
-        currentMainImage.unsetMainImage();
-        newMainImage.setMainImage();
+        jobPostImageDomainService.changeMainImage(currentImageId, newImageId);
     }
 }
